Squash from feature/arch-revamp
Former-commit-id: eed6ca81afcfe9cfbeda647c14ea0a61db8c7000
Former-commit-id: 922a817009a683fa3f4a6f75ba678ae23c9f9383

+++ DeviceIntegration/Scales/ScaleMeasurementData.cs
@@ -4,9 +4,9 @@
 	{
 		public MeasurementStatus Status { get; }
 
-		public long WeightGr { get; }
+		public double WeightGr { get; }
 
-		public ScaleMeasurementData(MeasurementStatus status, long weightGr)
+		public ScaleMeasurementData(MeasurementStatus status, double weightGr)
 		{
 			Status = status;
 			WeightGr = weightGr;
